pyproject.toml
[tool.poetry]
name = "infra-cluster-netangels"
version = "0.1.0"
description = "Nomad Cluster Deploy on NetAngels VDS (w/o Terraform)"
authors = ["justgithubaccount <kulikovyevgeny@outlook.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
